requirements.txt
opencv-python
ffmpeg-python
numpy
scikit-image